Solved Problem with Mocks in Hero.displayMove Test
When i mock the position it affects the functions that are normally available for positions. This means i need to call the function with the real position and then use the methods getX( ) and getY( ) to get the values of x and y for the mocked position.
This affected some of the tests that were based on the functions i altered, since the mocks stopped working in the way they were supposed to since the order changed.